chore: extract Options to CleanOptions
